pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "porntrex_api"
version = "1.0"
description = "A Python API for the Porn Site xvideos.com"
readme = "README.md"
requires-python = ">=3.6"
license = { text = "LGPLv3" }
authors = [
    { name = "Johannes Habel", email = "EchterAlsFake@proton.me" }
]
dependencies = [
    "bs4",
    "eaf_base_api",
    "httpx",
    "lxml",
    "h2",
    "brotli",
    "json5"
]
optional-dependencies = {av = ["av"], ffmpeg = ["ffmpeg-progress-yield"]}
classifiers = [
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python",
]

[project.urls]
Homepage = "https://github.com/EchterAlsFake/porntrex_api"

[tool.setuptools]
packages = { find = {} }